Modules
Modules to be used in this programme are datetime , os , smptplib , webbrowser , pyttsx3 (text-to-speech conversion library. it works offline) , speech recognition as sr , wikipedia , pyaudio , wave , gtts.
NOTE :- In pyttsx3 we will use sapi5 (The Speech Application Programming Interface or SAPI is an API developed by Microsoft to allow the use of speech recognition and speech synthesis within Windows applications).